requirements.txt
django
pillow
psycopg2-binary
python-decouple
djangorestframework
dj-rest-auth